Updating SLS QAQC
Created Outlier and NAFlag columns for all queries. Made yearOfInterest a bit more specific to take care of the situation when it's December but the current year's survey has not started yet.

Created an Outliers folder in the data-raw folder. SLS folder created specific in that folder. This is where the outlier excel sheet will be held. Changed naming convention of this file to also have today's date during the print.