Updating SLS code to incorporate in UDrive
Code now takes file location of SLS database from the U drive folder. Can still pull from the ftp website, but that needs to be specified.

The amount of tables available in the UDrive variant is a lot. Therefore, additional arguments provided in the readAccess.R file to accommodate.

Exported CSVs and RDS updated