Fixing float issues SLS.R
Added code to fix float issues caused by setting a column to field size "single" in Access and then it being read into R as "double" by default. This occurs when exporting to Excel also. This is simply a limitation of computer arithmetic but will address here by rounded to the 7th digits, which is one main mechanistic difference between "single" and "double" field size.

Updated the CSVs/RDS accordingly.