Added QAQC joining protocols and more annotations
QAQC of the joining process essentially compares the relational tables before joining and if we can recreate those relational tables after joining. If so, the joining did not duplicate/delete data; if not, the duplications and deletions must be accounted for.

Added more annotations in an attempt to make this code as user friendly as possible.