Updating before EDI publication
Added query checking code in the QAQC scripts to make sure that the recreations are 100% the same as the original in Access itself. This is the technical coding section in that script; it is currently not meant to be easily ran by the lead ES and only meant to check if my coding is right (after this initial check, I can't see a reason to run it again if the goal is to check to see if the original coding has been replicated).

Created a integrate code specific to EDI that's not written as a function to be published with the dataset.

Added additional console dialogs to help track the process.

NOTE: the relational tables are limited to only the 2021 season.

1. Fixed weird Unicode error
Changed Date from POSixct format to Date. Time and DateTime columns time zones fixed to be PST/PDT
2. Cleaned up integration code. Now, general workflow is 1) read in the data, 2) manipulation the individual tables, 3) join the tables, 3a) check after each join to find duplications/deletions, 4) create final join, 5) check again for duplications/deletions to ensure that database structure and keys are correctly assigned.