20 mm EDI synthesis code
Sculpting Lara's code to match the way that the SLS is currently coded; this is because I'd like to keep things consistent for the survey folks as they are learning through how to transition to R. There is nothing wrong with Lara's code.

Matched most things up. See lines 21-29 for comments about it