64bit office integration
Removed the download option from the 20mm.R script. I don't think we need to download from the FTP site anymore, as we are working from the U drive anyways. This code can be found in previous commits if required in the future. This removed the "bypass" argument from the readAccess script.

Integrated an option to state that you have 64 bit office. Before, the default was to always assume you had 32 bit office and always rely on 32 bit R. However, I now have 64 bit office and this future-proofs the time when 32-bit becomes obsolete for CDFW as well.