Initial QAQC script
Finished requested QAQC queries into R. Added GPS and cdec functionalities.

Next, need to prepare the scripts to export the relational tables next (manipulation steps). This is a departure from the 20 mm and SLS pipeline. I did not like the fact that you had to back trace within the pipeline to create the relational tables.