SKT first draft
Updated the workflow a bit. There is now an SKT folder in the data-raw folder that will host more than just the relational and joined tables for SKT. Within that folder, there is a folder named "filesToUpload" that will host the files that will be uploaded to EDI.

Relational Schema: new code to draw this schema in the metadata file itself. Requires the creation of the png file before being inputted into the markdown file itself.

Metadata_SKELETON file. First draft of the deep metadata skeleton. The aim is to have surveys document any sort of changes that may affect how one would interpret the fish data as a user.