SLS instructions and metadata updates
Including an instruction document to introduce users on how to use this workflow.

Updated metadata to be more in line with SKT. Still needs a bit of work. Awaiting Vanessa to update changes to protocols for 2023 season

Checked all code so far. Good so far. cdecHelper script, the right join into the pullCDEC df needing 'relationship = "many-to-many"' is indeed required.